Raster Splitter Rename files and algo from RastersplitterRP to Rastersplitter and Datasetsplitter Sum to dataset maker
