+++ enmapbox/apps/SpecDeepMap/__init__.py
@@ -1,7 +1,7 @@
 from enmapbox.gui.applications import EnMAPBoxApplication
 
 try:
-    from enmapbox.apps.SpecDeepMap.processingalgorithmDS_SUM12 import DatasetSplitter_SUM
+    from enmapbox.apps.SpecDeepMap.processingalgorithm_dataset_maker import DatasetMaker
     from enmapbox.apps.SpecDeepMap.processingalgorithm_raster_splitter import RasterSplitter
     from enmapbox.apps.SpecDeepMap.processingalgorithm_DL_UNET50_MOD_15_059_16 import DL_Train_MOD
     from enmapbox.apps.SpecDeepMap.processingalgorithm_PRED_GT_NO_DATA_mod11 import DL_Mapper
@@ -32,4 +32,4 @@ class SpecDeepMap(EnMAPBoxApplication):
         if wrongEnv:
             return []
         else:
-            return [RasterSplitterRP(), DatasetSplitter_SUM(), DL_Train_MOD(), DL_Mapper(), Tensorboard(),DL_Tester()]  # ,DL_Train_MOD()] #DL_Train()#DatasetSplitter() #,,DatasetSplitter(),DL_Train(),RasterSplitterR(),DatasetSplitter(),RasterSplitterRP()
+            return [RasterSplitter(), DatasetMaker(), DL_Train_MOD(), DL_Mapper(), Tensorboard(),DL_Tester()]  # ,DL_Train_MOD()] #DL_Train()#DatasetSplitter() #,,DatasetSplitter(),DL_Train(),RasterSplitterR(),DatasetSplitter(),RasterSplitterRP()
